[FEAT/#293] 토큰 재발급 auth로 이동
