+++ app/src/main/java/com/spoony/spoony/data/di/DataSourceModule.kt
@@ -9,7 +9,6 @@ import com.spoony.spoony.data.datasource.PostRemoteDataSource
 import com.spoony.spoony.data.datasource.ReportDataSource
 import com.spoony.spoony.data.datasource.ReviewRemoteDataSource
 import com.spoony.spoony.data.datasource.SpoonDataSource
-import com.spoony.spoony.data.datasource.TokenRefreshDataSource
 import com.spoony.spoony.data.datasource.UserRemoteDataSource
 import com.spoony.spoony.data.datasourceimpl.AuthRemoteDataSourceImpl
 import com.spoony.spoony.data.datasourceimpl.CategoryDataSourceImpl
@@ -20,7 +19,6 @@ import com.spoony.spoony.data.datasourceimpl.PostRemoteDataSourceImpl
 import com.spoony.spoony.data.datasourceimpl.ReportDataSourceImpl
 import com.spoony.spoony.data.datasourceimpl.ReviewRemoteDataSourceImpl
 import com.spoony.spoony.data.datasourceimpl.SpoonDataSourceImpl
-import com.spoony.spoony.data.datasourceimpl.TokenRefreshDataSourceImpl
 import com.spoony.spoony.data.datasourceimpl.UserRemoteDataSourceImpl
 import dagger.Binds
 import dagger.Module
@@ -52,12 +50,6 @@ abstract class DataSourceModule {
     abstract fun bindAuthRemoteDataSource(
         authRemoteDataSourceImpl: AuthRemoteDataSourceImpl
     ): AuthRemoteDataSource
-
-    @Binds
-    @Singleton
-    abstract fun bindTokenRefreshDataSource(
-        tokenRefreshDataSourceImpl: TokenRefreshDataSourceImpl
-    ): TokenRefreshDataSource
 
     @Binds
     @Singleton
