+++ app/src/main/java/com/spoony/spoony/data/service/AuthService.kt
@@ -5,6 +5,7 @@ import com.spoony.spoony.data.dto.request.SignInRequestDto
 import com.spoony.spoony.data.dto.request.SignUpRequestDto
 import com.spoony.spoony.data.dto.response.SignInResponseDto
 import com.spoony.spoony.data.dto.response.SignUpResponseDto
+import com.spoony.spoony.data.dto.response.TokenResponseDto
 import retrofit2.http.Body
 import retrofit2.http.Header
 import retrofit2.http.POST
@@ -21,4 +22,9 @@ interface AuthService {
         @Header("Authorization") authorization: String,
         @Body signUpRequestDto: SignUpRequestDto
     ): BaseResponse<SignUpResponseDto>
+
+    @POST("api/v1/auth/refresh")
+    suspend fun refreshToken(
+        @Header("Authorization") authorization: String
+    ): BaseResponse<TokenResponseDto>
 }
