--- app/src/main/java/com/spoony/spoony/domain/repository/TokenRefreshRepository.kt
@@ -1,7 +1,0 @@
-package com.spoony.spoony.domain.repository
-
-import com.spoony.spoony.domain.entity.TokenEntity
-
-interface TokenRefreshRepository {
-    suspend fun refreshToken(token: String): Result<TokenEntity>
-}
